isolated raw_data functions from data.py to raw_data.py; and yf/kaggle functions to own modules
