+++ main_training.py
@@ -13,14 +13,16 @@ from data import (
     build_features,
     clean_raw_data,
     create_X_y_multistep,
-    load_raw_data,
-    remove_raw_data,
-    sample_tickers_dates,
     split_train_test_panel,
 )
 from gcp_functions import upload_file_to_folder
 from load_configs import Configs
 from models import create_fit_xgbregressor_chain, evaluate_all
+from raw_data import (
+    load_raw_data,
+    remove_raw_data,
+    sample_tickers_dates,
+)
 from utils import get_pipreqs_from_pyproject
 
 # Global parameters
